+++ app/plex_auto_collections.py
@@ -143,7 +143,7 @@ def update_from_config(config_path, plex, headless=False, no_meta=False, no_imag
         "tvdb_show",
         "imdb_list",
         "trakt_list",
-        "trakt_trending",
+        "trakt_watchlist",
         "tautulli"
     ]
     plex_searches = [
@@ -426,6 +426,8 @@ def update_from_config(config_path, plex, headless=False, no_meta=False, no_imag
                     methods.append((method_name, get_attribute_list(collections[c][m])))
                 elif method_name == "trakt_trending":
                     methods.append((method_name, [regex_first_int(collections[c][m], method_name, default=30)]))
+                elif method_name == "trakt_watchlist":
+                    methods.append((method_name, get_attribute_list(collections[c][m])))
                 elif method_name == "tautulli": #TODO:test
                     try:
                         new_dictionary = {}
@@ -465,7 +467,7 @@ def update_from_config(config_path, plex, headless=False, no_meta=False, no_imag
                 if missing:
                     if libtype == "movie":
                         method_name = "IMDb" if "imdb" in m else "Trakt" if "trakt" in m else "TMDb"
-                        if m in ["trakt_list", "tmdb_list"]:
+                        if m in ["trakt_list", "trakt_watchlist", "tmdb_list"]:
                             print("| {} missing movie{} from {} List: {}".format(len(missing), "s" if len(missing) > 1 else "", method_name, v))
                         elif m == "imdb_list":
                             print("| {} missing movie{} from {} List: {}".format(len(missing), "s" if len(missing) > 1 else "", method_name, v[0]))
@@ -484,7 +486,7 @@ def update_from_config(config_path, plex, headless=False, no_meta=False, no_imag
                                 add_to_radarr(config_path, missing)
                     elif libtype == "show":
                         method_name = "Trakt" if "trakt" in m else "TVDb" if "tvdb" in m else "TMDb"
-                        if m in ["trakt_list", "tmdb_list"]:
+                        if m in ["trakt_list", "trakt_watchlist", "tmdb_list"]:
                             print("| {} missing show{} from {} List: {}".format(len(missing), "s" if len(missing) > 1 else "", method_name, v))
                         elif m == "trakt_trending":
                             print("| {} missing show{} from {} List: Trending (top {})".format(len(missing), "s" if len(missing) > 1 else "", method_name, v))